fix(review): colocate session index reader tests
